Migrate Tecan Infinite 200 PRO to Device/Driver/Capability architecture
Split the monolithic ExperimentalTecanInfinite200ProBackend (~1345 lines)
into the new architecture:

- pylabrobot/tecan/infinite/protocol.py: pure protocol utilities (framing,
  stream parsing, decoders, calibration math)
- pylabrobot/tecan/infinite/driver.py: TecanInfiniteDriver (USB I/O,
  scan orchestration, well-to-stage geometry)
- pylabrobot/tecan/infinite/absorbance_backend.py: AbsorbanceBackend impl
- pylabrobot/tecan/infinite/fluorescence_backend.py: FluorescenceBackend impl
- pylabrobot/tecan/infinite/luminescence_backend.py: LuminescenceBackend impl
- pylabrobot/tecan/infinite/infinite.py: TecanInfinite200Pro device frontend

The legacy wrapper now delegates to the new backends (zero duplicated
logic). Legacy tests patched to target the driver directly. All 30 tests
pass (18 legacy + 12 new).

Also updates the docs notebook to use the new API and adds Sphinx API
docs for pylabrobot.tecan.
