+++ docs/_static/devices.json
@@ -654,7 +654,7 @@
       "luminescence"
     ],
     "status": "Mostly",
-    "docs": "/user_guide/02_analytical/plate-reading/tecan-infinite.html",
+    "docs": "/user_guide/tecan/infinite/hello-world.html",
     "oem": "https://lifesciences.tecan.com/infinite-200-pro"
   },
   {
