+++ docs/api/pylabrobot.rst
@@ -44,4 +44,5 @@ Manufacturers
     pylabrobot.molecular_devices
     pylabrobot.opentrons
     pylabrobot.qinstruments
+    pylabrobot.tecan
     pylabrobot.thermo_fisher
