+++ docs/api/pylabrobot.tecan.rst
@@ -1,0 +1,55 @@
+.. currentmodule:: pylabrobot.tecan
+
+pylabrobot.tecan package
+========================
+
+Infinite 200 PRO
+-----------------
+
+.. currentmodule:: pylabrobot.tecan.infinite.infinite
+
+.. autosummary::
+  :toctree: _autosummary
+  :nosignatures:
+  :recursive:
+
+    TecanInfinite200Pro
+
+.. currentmodule:: pylabrobot.tecan.infinite.driver
+
+.. autosummary::
+  :toctree: _autosummary
+  :nosignatures:
+  :recursive:
+
+    TecanInfiniteDriver
+
+.. currentmodule:: pylabrobot.tecan.infinite.absorbance_backend
+
+.. autosummary::
+  :toctree: _autosummary
+  :nosignatures:
+  :recursive:
+
+    TecanInfiniteAbsorbanceBackend
+    TecanInfiniteAbsorbanceParams
+
+.. currentmodule:: pylabrobot.tecan.infinite.fluorescence_backend
+
+.. autosummary::
+  :toctree: _autosummary
+  :nosignatures:
+  :recursive:
+
+    TecanInfiniteFluorescenceBackend
+    TecanInfiniteFluorescenceParams
+
+.. currentmodule:: pylabrobot.tecan.infinite.luminescence_backend
+
+.. autosummary::
+  :toctree: _autosummary
+  :nosignatures:
+  :recursive:
+
+    TecanInfiniteLuminescenceBackend
+    TecanInfiniteLuminescenceParams
