+++ docs/user_guide/tecan/index.md
@@ -1,0 +1,7 @@
+# Tecan
+
+```{toctree}
+:maxdepth: 1
+
+infinite/hello-world
+```
